+++ src/test/java/com/clickhouse/kafka/connect/sink/ClickHouseSinkTaskWithSchemaProxyTest.java
@@ -6,6 +6,7 @@ import com.clickhouse.kafka.connect.ClickHouseSinkConnector;
 import com.clickhouse.kafka.connect.sink.db.helper.ClickHouseHelperClient;
 import com.clickhouse.kafka.connect.sink.helper.ClickHouseTestHelpers;
 import com.clickhouse.kafka.connect.sink.helper.SchemaTestData;
+import com.clickhouse.kafka.connect.util.Utils;
 import eu.rekawek.toxiproxy.Proxy;
 import eu.rekawek.toxiproxy.ToxiproxyClient;
 import org.apache.kafka.connect.errors.DataException;
@@ -225,7 +226,11 @@ public class ClickHouseSinkTaskWithSchemaProxyTest {
 
         ClickHouseSinkTask chst = new ClickHouseSinkTask();
         chst.start(props);
-        assertThrows(DataException.class, () -> chst.put(sr), "Did not detect wrong date conversion ");
+        try {
+            chst.put(sr);
+        } catch (RuntimeException e) {
+            assertTrue(Utils.getRootCause(e) instanceof DataException, "Did not detect wrong date conversion ");
+        }
         chst.stop();
     }